requirements.txt
fastapi
uvicorn[standard] # [standard] includes httptools and other performance enhancements
sqlalchemy[asyncio]
asyncpg # For async PostgreSQL connection
psycopg2-binary # Fallback/optional for direct sync connections, but asyncpg is preferred with FastAPI
python-multipart # For file uploads
python-dotenv # For loading .env files
qdrant-client
sentence-transformers # For embeddings
pydantic-settings # For structured settings management
aiofiles # For async file operations
python-magic # To detect file types (optional, but good for robust file handling)
boto3 # AWS SDK for S3
# python-zipstream # For streaming zip content, if needed (or just use tempfile and shutil)
# unzip-http is not a standard PyPI package, will rely on python's zipfile or external tools.